Completed showing the posts. Note: this version doesn't fully work but I'm leaving it for now (see below).
These is some mess with the way the properties are defined in this tutorial.
However, since this is enough for what I need right now (i.e. application design steps), I'm leaving it for now. Maybe in the future I will refactor it and make it work.
Maybe I can also use the actual services (instead of simulations) in the future.